Modified the percentValue to Integer

+++ spring-jpa-postgres-master/src/main/java/com/springboot/jpa/postgres/rest/model/InductsForLastHourOutput.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class InductsForLastHourOutput{
 	private List<InductsForLastHour> inductsForLastHour;
 
-	private double percentValue;
+	private Integer percentValue;
 	
 	private Integer currentValue;
